+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 2.0.1 (TBD)
 
-* TBD
+* Geometry Projection Transformations for Curve Polygons and derived Geometry types
 
 ## [2.0.0](https://github.com/ngageoint/simple-features-proj-java/releases/tag/2.0.0) (05-17-2018)